@@ -1,25 +1,28 @@
 name: mock_creator_generator
 description: Creates a mock for classes with a single call method making it easy for us to create those mocks automatically.
-version: 1.0.14-nullsafety4
+version: 1.0.16
 homepage: https://github.com/atreeon/MockCreator
 publish_to: none
 
 environment:
-  sdk: ">=2.7.2 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
-  analyzer: '>=1.7.0 <=3.0.0'
-  build: ^2.1.0
+  analyzer: '>=1.7.0 <=4.4.0'
+  build: ^2.1.1
   source_gen: ^1.1.1
-#  analyzer_models: ^1.0.3
-#  generator_common: ^1.0.0-nullsafety7
   generator_common:
-    path: ../../generator_common
-#  1.0.0-nullsafety2
+    git:
+      url: https://github.com/atreeon/generator_common
+      ref: master
   meta: ^1.1.8
   mock_creator_annotation:
     path: ../mock_creator_annotation
 
+dependency_overrides:
+  mock_creator_annotation:
+    path: ../mock_creator_annotation
+
 dev_dependencies:
-  build_runner: ^2.1.4
-  test: ^1.18.2
+  test: ^1.24.3
+  build_runner: ^2.3.3
